depricate otto v1 functions
